add Input module
