+++ Project.toml
@@ -4,6 +4,8 @@ authors = ["CliMA Contributors <clima-software@caltech.edu>"]
 version = "0.1.2"
 
 [deps]
+ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
+ClimaAtmos = "b2c96348-7fb7-4fe0-8da9-78d88439e717"
 ClimaComms = "3a4d1b5c-c61d-41fd-a00a-5873ba7a1b0d"
 ClimaCore = "d414da3d-4745-48bb-8d80-42e94e092884"
 ClimaUtilities = "b3f4f4ca-9299-4f7f-bd9b-81e1242a7513"
@@ -14,8 +16,11 @@ SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 SurfaceFluxes = "49b00bb7-8bd4-4f2b-b78c-51cd0450215f"
 Thermodynamics = "b60c26fb-14c3-4610-9d3e-2d17fe7ff00c"
+YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 
 [compat]
+ArgParse = "1"
+ClimaAtmos = "0.27, 0.28, 0.29, 0.30, 0.31"
 ClimaComms = "0.6.2"
 ClimaCore = "0.14.25"
 ClimaUtilities = "0.1.22"
@@ -26,4 +31,5 @@ SciMLBase = "~2.110, ~2.111, ~2.112"
 StaticArrays = "1.9.8"
 SurfaceFluxes = "0.12.3, 0.13, 0.14"
 Thermodynamics = "0.14.1, 0.15"
+YAML = "0.4"
 julia = "1.10"
