+++ experiments/calibration/model_interface.jl
@@ -5,7 +5,7 @@ include(joinpath(pkgdir(ClimaCoupler), "experiments", "ClimaEarth", "setup_run.j
 function ClimaCalibrate.forward_model(iter, member)
     config_file =
         joinpath(pkgdir(ClimaCoupler), "experiments", "calibration", "model_config.yml")
-    config_dict = get_coupler_config_dict(config_file)
+    config_dict = Input.get_coupler_config_dict(config_file)
 
     # Run for a shorter time if SHORT_RUN is set
     if SHORT_RUN
