+++ src/ClimaCoupler.jl
@@ -12,5 +12,6 @@ include("ConservationChecker.jl")
 include("FluxCalculator.jl")
 include("FieldExchanger.jl")
 include("Checkpointer.jl")
+include("Input.jl")
 
 end
